+++ src/main/java/br/wk/projeto/bancosangue/dto/TipoSanguineoDTO.java
@@ -1,7 +1,5 @@
 package br.wk.projeto.bancosangue.dto;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,7 +12,7 @@ import java.math.BigInteger;
 
 public class TipoSanguineoDTO {
 
-    private BigInteger id;
+    private Long id;
     private String codigo;
     private String doar;
     private String receber;
