+++ src/main/java/br/wk/projeto/bancosangue/model/TipoSanguineo.java
@@ -1,10 +1,8 @@
 package br.wk.projeto.bancosangue.model;
 
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.annotation.CreatedBy;
 
 import java.math.BigInteger;
 
@@ -16,7 +14,7 @@ public class TipoSanguineo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private BigInteger id;
+    private Long id;
     @Column(name = "codigo", nullable = false, length = 255)
     private String codigo;
     @Column(name = "doar", nullable = false, length = 255)
