+++ src/main/java/br/wk/projeto/bancosangue/repository/TipoSanguineoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface TipoSanguineoRepository extends JpaRepository<TipoSanguineo, BigInteger> {
+public interface TipoSanguineoRepository extends JpaRepository<TipoSanguineo, Long> {
 }
